back-end/requirements.txt
Django==1.11.27
djangorestframework==3.9.4
django-cors-headers==3.0.2
psycopg2
djangorestframework-jwt
cryptography
python-jose
uWSGI==2.0.17.1
